requirements.txt
streamlit==1.46.1
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
tqdm==4.67.1
